* basic linux compilation script using cmake, thanks to Joe LeVeque * modified script to include theora asm optimizations and glut_player executable * various linux compilation fixes * defaulted to 32/64 bit build in Mac throug xcconfig files

+++ theoraplayer/src/AVFoundation/TheoraVideoClip_AVFoundation.mm
@@ -171,7 +171,7 @@ bool TheoraVideoClip_AVFoundation::decodeNextFrame()
 			CVPixelBufferLockBaseAddress(imageBuffer, 0);
 			void *baseAddress = CVPixelBufferGetBaseAddress(imageBuffer);
 			
-			mStride = CVPixelBufferGetBytesPerRow(imageBuffer);
+			mStride = (int) CVPixelBufferGetBytesPerRow(imageBuffer);
 			size_t width = CVPixelBufferGetWidth(imageBuffer);
 			size_t height = CVPixelBufferGetHeight(imageBuffer);
